Crate android_ffi

Structs

Motion
Data about a motion event.

Enums

AssetError

Event
An event triggered by the Android environment.

MotionAction
The type of pointer action in a motion event.

Statics

ANDROID_APP
This static variable will store the android_app* on creation, and set it back to 0 at destruction. Apart from this, the static is never written, so there is no risk of race condition.

Traits

SyncEventHandler

Functions

add_sender
Adds a sender where events will be sent to.

add_sender_missing
Adds a sender where events will be sent to, but also sends any missing events to the sender object.

add_sync_event_handler
Adds a SyncEventHandler which will process sync events from the polling loop.

cargo_apk_injected_glue_add_sender⚠

cargo_apk_injected_glue_add_sender_missing⚠

cargo_apk_injected_glue_add_sync_event_handler⚠

cargo_apk_injected_glue_get_native_window⚠

cargo_apk_injected_glue_load_asset⚠

cargo_apk_injected_glue_remove_sync_event_handler⚠

cargo_apk_injected_glue_set_multitouch⚠

cargo_apk_injected_glue_wake_event_loop⚠

cargo_apk_injected_glue_write_log⚠

get_app
Return a reference to the application structure.

get_native_window⚠
Returns a handle to the native window.

inputs_callback
The callback for input.

load_asset

remove_sync_event_handler
Removes a SyncEventHandler.

set_multitouch

write_log

Type Definitions

pthread_attr_t

pthread_mutexattr_t

pthread_t